+++ testing/props.test.js
@@ -1,10 +1,19 @@
-let StateManagement = require("../test/StateManagement").default;
+const StateManagement = require("../test/StateManagement").default;
 
-let prop = "<h1>Hello {name - prop}!</h1>";
-let managerState = new StateManagement();
+const propInBars = "<h1>Hello {name - prop}!</h1>";
+const propInBarManager = new StateManagement();
 
-managerState.getHTMLString(prop);
+propInBarManager.getHTMLString(propInBars);
 
-test("test get prop from template", ()=>{
-    expect(managerState.props).toContain("name");
+test("test get prop from bars", ()=>{
+    expect(propInBarManager.props).toContain("name");
+});
+
+const propInBindAttributes = "<h1 :class='color - prop'>Hello World!</h1>";
+const propInBindAttributesManager = new StateManagement();
+
+propInBindAttributesManager.getHTMLString(propInBars);
+
+test("test get prop from bind attribute", ()=>{
+    expect(propInBindAttributesManager.props).toContain("color");
 });